split di module package.
