+++ app/src/main/java/com/kazakago/cleanarchitecture/CleanApplication.kt
@@ -4,15 +4,15 @@ import android.app.Application
 import com.github.salomonbrys.kodein.Kodein
 import com.github.salomonbrys.kodein.KodeinAware
 import com.github.salomonbrys.kodein.lazy
-import com.kazakago.cleanarchitecture.di.applicationModule
-import com.kazakago.cleanarchitecture.di.dataModule
-import com.kazakago.cleanarchitecture.di.domainModule
-import com.kazakago.cleanarchitecture.di.webModule
+import com.kazakago.cleanarchitecture.data.di.dataModule
+import com.kazakago.cleanarchitecture.domain.di.domainModule
+import com.kazakago.cleanarchitecture.presentation.di.presentationModule
+import com.kazakago.cleanarchitecture.web.di.webModule
 
 open class CleanApplication : Application(), KodeinAware {
 
     override val kodein: Kodein by Kodein.lazy {
-        import(applicationModule(this@CleanApplication))
+        import(presentationModule(this@CleanApplication))
         import(domainModule())
         import(dataModule())
         import(webModule())
