+++ data/src/main/java/com/kazakago/cleanarchitecture/data/di/DataModule.kt
@@ -1,4 +1,4 @@
-package com.kazakago.cleanarchitecture.di
+package com.kazakago.cleanarchitecture.data.di
 
 import com.github.salomonbrys.kodein.Kodein
 import com.github.salomonbrys.kodein.bind
