+++ presentation/src/main/java/com/kazakago/cleanarchitecture/presentation/di/PresentationModule.kt
@@ -1,10 +1,10 @@
-package com.kazakago.cleanarchitecture.di
+package com.kazakago.cleanarchitecture.presentation.di
 
 import android.content.Context
 import com.github.salomonbrys.kodein.Kodein
 import com.github.salomonbrys.kodein.bind
 import com.github.salomonbrys.kodein.instance
 
-fun applicationModule(context: Context) = Kodein.Module {
+fun presentationModule(context: Context) = Kodein.Module {
     bind<Context>() with instance(context)
 }
