+++ web/build.gradle
@@ -27,6 +27,8 @@ dependencies {
     implementation project(':domain')
     //Kotlin
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+    //Kodein
+    implementation 'com.github.salomonbrys.kodein:kodein:4.1.0'
     //Rx
     implementation 'io.reactivex.rxjava2:rxjava:2.1.12'
     implementation 'io.reactivex.rxjava2:rxkotlin:2.2.0'
